Update API spec with currently published version
I pulled the latest spec from
https://api.ynab.com/papi/open_api_spec.yaml
and wrote it to `data/open_api_spec.yaml`. Unfortunately, the swagger
JSON file that was previously published at
https://github.com/ynab/ynab-sdk-js/blob/master/spec-v1-swagger.json
doesn't seem to be available anymore.

This commit, therefore, replaces the JSON spec with a YAML spec,
changes the regen-ynab-api command to reference that file instead of
the old one, and changes the extra readme to reference the new source of
the new spec.